Fix overload/hightemp conditions trigger loss of communication
Since Venus v2.0, vebus alarm paths have  been changed but
the script still checking the old ones. This triggers a loss
of communication state when overload or high temperature alarm
conditions are enabled.

This bug only affects when the multi is connected to vebus, when
conneted to CAN-bus works as expected.