Get readings from systemcalc instead of directly from the service
Currently the dbus-generator service gets the system configured
battery service instance from systemcalc and then it reads the
voltage, current and SOC values directly from the service.

This commit makes dbus-generator read the battery values from
systemcalc. Systemcalc provides calculated values so in case
of a system without a battery monitor but with a Multi,
systemcalc takes into account the PV current instead assuming
that all the current is drawn from the battery.