publish accumulated runtime on dbus service
This value is stored in localsettings and has always been available
from there, but in localsettings it is stored according to the
driver that is in use, either relay or fischerpanda.

To remove the distinction between genetic relay-controlled and
fischerpanda, this is now mirrored on the service itself. That will
allow a single attribute logged to VRM to account for the generator
runtime, instead of two.