support AcIn1Available readout from multi/quattro
When the generator is on AC-in-2, the existing implementation would
wait for AC1 to become active, which a Quattro will do automatically
when the grid returns on AC-in-1, and then stop the generator. But it
would not stop the generator if AC-in-1 is not accepted, which could
potentially leave the generator running without being used at all.

By using the readout support, it will at least turn off the generator,
unless another condition keeps it running.

This is backwards compatible. If the readout support is not in the
Quattro firmware, the old behaviour is kept.