cooldown: add 15 second delay for engine to stop
It turns out that some generators takes a long time
to respond to the instruction to switch off. When cooldown is
configured, the Multi is instructed to ignore AC-in while the
generator cools down, and AC-in is un-ignored once the generator
is stopped.

If the generator takes too long to stop, however, AC-in may
be re-engaged before the generator stops. To avoid this,
there is now a 15 second delay to allow the generator
to stop cleanly.

https://github.com/victronenergy/venus/issues/1031