avoid race condition on genset/AutoStart
On some gensets, the /AutoStart path may not be valid yet at the
time .genset service comes up. Instead of requiring this value
to be valid, rather check if the path exists. Services that
cannot autostart, such as when using an energy meter to measure
a genset, will not have the /AutoStart path.