warmup-cooldown: Only control generator AC inputs
In the past, we would ignore AC in for warmup and cooldown
only on the inputs marked as generator, but when we un-ignored
them we would do it for all interfaces. This would however cause
dbus-generator to conflict with other systems that need to ignore
AC-in when the input is grid (the load shedding implementation).

Therefore only ignore and un-ignore AC-in if it is marked as
generator.