Helper relay for connected gensets
Inteded for connected gensets which cannot be started and stopped
over Modbus. These gensets do not report a valid '/Start' path.
If this is the case, a helper relay must be configured which start
and stops the genset.

For connected gensets which do support start and stop commands over
Modbus, the helper relay can still be configured and will be enabled/disabled
in conjunction with the genset (which is controlled with '/Start').

The option 5:"Connected genset helper relay" is added to the gui for this.

https://github.com/victronenergy/venus/issues/1286