Fix minor UI issue when stopping generator
When startstop stops, it passed through 'cool-down', to 'stopping' and to 'stopped'.
However, the GUI's show the running by condition value if its nonzero when startstop is in 'stopped'.
This results in e.g. 'Started manually' being shown very briefly inbetween 'stopping' and 'stopped',
because dbus-generator clears '/State' before clearing '/RunningByConditionCode'.
The solution is to reverse the order.