Handle 'nobattery' battery monitor setting
It used to be possible to set the battery monitor to be used for the start/stop conditions.
This setting is deprecated and set to 'default' by default'. If set to 'default',
the GUI won't show the setting.

However, there are still systems which do have a value other than 'default'.
The GUI then also presents the option 'No battery monitor'.

dbus-generator-starter used to crash when that option was selected. This commit fixes that.